initial work to convert to LWRP DSL
