+++ Guardfile
@@ -15,9 +15,9 @@ guard :rspec, :cmd => 'chef exec /opt/chefdk/embedded/bin/rspec', :all_on_start 
   watch('spec/spec_helper.rb')      { 'spec' }
 end
 
-guard :rake, :task => 'spec', :all_on_start => false, :notification => false do
-  watch(%r{^libraries\/(.+)\.rb$})
-  watch(%r{^spec\/(.+)_spec\.rb$})
-  watch(%r{^(recipes)\/(.+)\.rb$}) { |m| "spec/#{m[1]}_spec.rb" }
-  watch('spec/spec_helper.rb') { 'spec' }
-end
+# guard :rake, :task => 'spec', :all_on_start => false, :notification => false do
+#   watch(%r{^libraries\/(.+)\.rb$})
+#   watch(%r{^spec\/(.+)_spec\.rb$})
+#   watch(%r{^(recipes)\/(.+)\.rb$}) { |m| "spec/#{m[1]}_spec.rb" }
+#   watch('spec/spec_helper.rb') { 'spec' }
+# end
