+++ libraries/helpers.rb
@@ -18,10 +18,13 @@
 # limitations under the License.
 #
 
-require 'chef/mixin/shell_out'
-include Chef::Mixin::ShellOut
-module Runit
+module RunitCookbook
   module Helpers
+    def inside_docker?
+      results = `cat /proc/1/cgroup`.strip.split("\n")
+      results.any?{|val| /docker/ =~ val}
+    end
+    
     def runit_installed?
       return true if runit_rpm_installed? || (runit_executable? && runit_sv_works?)
     end
@@ -38,8 +41,60 @@ module Runit
     def runit_rpm_installed?
       shell_out('rpm -qa | grep -q "^runit"').exitstatus == 0
     end
+
+    def runit_send_signal(signal, friendly_name = nil)
+      friendly_name ||= signal
+      converge_by("send #{friendly_name} to #{new_resource}") do
+        shell_out!("#{new_resource.sv_bin} #{sv_args}#{signal} #{service_dir_name}")
+        Chef::Log.info("#{new_resource} sent #{friendly_name}")
+      end
+    end
+
+    def running?
+      cmd = shell_out("#{new_resource.sv_bin} #{sv_args}status #{service_dir_name}")
+      (cmd.stdout =~ /^run:/ && cmd.exitstatus == 0)
+    end
+
+    def log_running?
+      cmd = shell_out("#{new_resource.sv_bin} #{sv_args}status #{service_dir_name}/log")
+      (cmd.stdout =~ /^run:/ && cmd.exitstatus == 0)
+    end
+
+    def enabled?
+      ::File.exists?(::File.join(service_dir_name, 'run'))
+    end
+
+    def log_service_name
+      ::File.join(new_resource.service_name, 'log')
+    end
+
+    def sv_dir_name
+      ::File.join(new_resource.sv_dir, new_resource.service_name)
+    end
+
+    def sv_args
+      sv_args = ''
+      sv_args += "-w '#{new_resource.sv_timeout}' " unless new_resource.sv_timeout.nil?
+      sv_args += '-v ' if new_resource.sv_verbose
+      sv_args
+    end
+
+    def service_dir_name
+      ::File.join(new_resource.service_dir, new_resource.service_name)
+    end
+
+    def log_dir_name
+      ::File.join(new_resource.service_dir, new_resource.service_name, log)
+    end
+
+    def template_cookbook
+      new_resource.cookbook.nil? ? new_resource.cookbook_name.to_s : new_resource.cookbook
+    end
+
+    def default_logger_content
+      "#!/bin/sh
+exec svlogd -tt /var/log/#{new_resource.service_name}"
+    end
+    
   end
 end
-
-Chef::Recipe.send(:include, Runit::Helpers)
-Chef::Resource.send(:include, Runit::Helpers)
